+++ src/TwentyI/Provider.php
@@ -71,9 +71,14 @@ class Provider extends SharedHosting implements ProviderInterface
             throw $this->errorResult('Domain name is required');
         }
 
+        if (empty($params->location)) {
+            throw $this->errorResult('Location missing');
+        }
+
         $hostingId = $this->api()->createPackage(
             $params->package_name,
             $params->domain,
+            $this->getLocationIdentifier($params->location),
             $customerId = $this->findOrCreateUser($params)
         );
 
@@ -252,6 +257,7 @@ class Provider extends SharedHosting implements ProviderInterface
             ->setPackageName((string)$packageName)
             ->setSuspended($suspended)
             ->setSuspendReason($suspendReason)
+            ->setLocation($info->web->info->zone)
             ->setIp($ip);
     }
 
@@ -399,5 +405,33 @@ class Provider extends SharedHosting implements ProviderInterface
             $this->configuration->general_api_key,
             $this->configuration->debug ? $this->getLogger() : null
         ));
+    }
+
+    /**
+     * Checks if the provided location string matches with any occurrence
+     * from 20i endpoint, if not notify the location si not valid.
+     *
+     * If the response from the API is empty then there was an error
+     * retrieving the locations list.
+     *
+     * @param string $location
+     * @return string
+     */
+    protected function getLocationIdentifier(string $location): string
+    {
+        $location = trim($location);
+        $locations = $this->api()->locations();
+
+        if (empty($locations)) {
+            throw $this->errorResult('There was an error retrieving the locations');
+        }
+
+        foreach ($locations as $key => $value) {
+            if ($location === trim($key) || $location === trim($value)) {
+                return $key;
+            }
+        }
+
+        throw $this->errorResult('Location value not valid');
     }
 }