+++ src/Enhance/Data/Configuration.php
@@ -15,6 +15,7 @@ use Upmind\ProvisionBase\Provider\DataSet\Rules;
  * @property-read string $access_token API access token
  * @property-read string|null $sso_destination Which control panel to log into, either Enhance or Wordpress
  * @property-read bool|null $ignore_ssl_errors When set to true, SSL will not be verified
+ * @property-read bool|null $create_subscription_only When enabled, create() creates a subscription without creating a website
  * @property-read bool $remove_www Whether or not to strip www. from domain names when creating new subscriptions
  * @property-read bool|null $debug Whether or not to enable debug logging of API requests
  */
@@ -28,6 +29,7 @@ class Configuration extends DataSet
             'access_token' => ['required', 'string'],
             'sso_destination' => ['in:enhance,wordpress'],
             'ignore_ssl_errors' => ['boolean'],
+            'create_subscription_only' => ['boolean'],
             'remove_www' => ['boolean'],
             'debug' => ['boolean'],
         ]);