+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the package will be documented in this file.
 
+## TBC
+- Updated 20i create() method to include the location.
+- Updated the getInfo() method to return the location.
+
 ## [v6.5.2](https://github.com/upmind-automation/provision-provider-shared-hosting/releases/tag/v6.5.2) - 2024-01-08
 
 - Update DirectAdmin add `ip_status` configuration value to specify IP status type